examples/torchscript/requirements.txt
clearml
torch
torchvision 
Pillow
nvidia-pyindex
tritonclient
requests